usb gadget: CAF 2.6.35 gadget

+++ include/linux/usb/f_mtp.h
@@ -29,7 +29,7 @@ struct mtp_file_range {
 	/* offset in file for start of transfer */
 	loff_t  	offset;
 	/* number of bytes to transfer */
-	size_t		length;
+	int64_t		length;
 };
 
 struct mtp_event {
@@ -51,3 +51,4 @@ struct mtp_event {
 #define MTP_SEND_EVENT             _IOW('M', 3, struct mtp_event)
 
 #endif /* __LINUX_USB_F_MTP_H */
+
